Extract shared code to library
