+++ AnonymousTokens.sln
@@ -9,6 +9,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Client-side", "Client-side"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Server-side", "Server-side", "{FC4D5FE2-898D-4C67-8094-FDB6E71247CD}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Library", "Library", "{4DDE2573-8947-4348-95B4-688102A8294D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AnonymousTokensShared", "src\Shared\AnonymousTokensShared\AnonymousTokensShared.csproj", "{D98EB88A-9922-4AA5-8C97-6206C7196CD5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -19,12 +23,17 @@ Global
 		{CCB7635D-7D0F-4371-AC14-828E208790AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CCB7635D-7D0F-4371-AC14-828E208790AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CCB7635D-7D0F-4371-AC14-828E208790AF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D98EB88A-9922-4AA5-8C97-6206C7196CD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D98EB88A-9922-4AA5-8C97-6206C7196CD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D98EB88A-9922-4AA5-8C97-6206C7196CD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D98EB88A-9922-4AA5-8C97-6206C7196CD5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{CCB7635D-7D0F-4371-AC14-828E208790AF} = {E4DAF051-5F2C-458A-8AD3-2E0B1E7D3C52}
+		{D98EB88A-9922-4AA5-8C97-6206C7196CD5} = {4DDE2573-8947-4348-95B4-688102A8294D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {EC41792F-67C9-4BD6-B598-E3E2ECDD924C}
