+++ src/Client/AnonymousTokensConsole/AnonymousTokensConsole.csproj
@@ -9,4 +9,8 @@
     <PackageReference Include="Portable.BouncyCastle" Version="1.8.8" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\Shared\AnonymousTokensShared\AnonymousTokensShared.csproj" />
+  </ItemGroup>
+
 </Project>
