+++ src/Client/AnonymousTokensConsole/Program.cs
@@ -1,4 +1,6 @@
-﻿using Org.BouncyCastle.Asn1.X9;
+﻿using AnonymousTokensShared;
+
+using Org.BouncyCastle.Asn1.X9;
 using Org.BouncyCastle.Crypto.Parameters;
 using Org.BouncyCastle.Math;
 using Org.BouncyCastle.Math.EC;
